python-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.3.3
pydantic-settings==2.6.1
jinja2==3.1.4
pytest==8.3.4
httpx==0.28.1
